+++ package.json
@@ -23,13 +23,14 @@
   "devDependencies": {
     "@babel/core": "7.18.5",
     "@babel/eslint-parser": "7.18.2",
+    "@dylandepass/helix-web-library": "^1.9.0",
+    "@esm-bundle/chai": "4.3.4-fix.0",
+    "@web/test-runner": "0.13.30",
+    "@web/test-runner-commands": "0.6.2",
     "chai": "4.3.6",
     "eslint": "8.17.0",
     "eslint-config-airbnb-base": "15.0.0",
     "eslint-plugin-import": "2.26.0",
-    "@esm-bundle/chai": "4.3.4-fix.0",
-    "@web/test-runner": "0.13.30",
-    "@web/test-runner-commands": "0.6.2",
     "sinon": "14.0.0",
     "stylelint": "14.9.1",
     "stylelint-config-prettier": "9.0.3",
