+++ scripts/delayed.js
@@ -1,5 +1,5 @@
 // eslint-disable-next-line import/no-cycle
-import { sampleRUM } from './scripts.js';
+import { sampleRUM } from './helix-web-library.esm.js';
 
 // Core Web Vitals RUM collection
 sampleRUM('cwv');
